refactor snake case file names to kebab case
